Fixed and modified io operations

+++ src/main/java/io/TableIO.java
@@ -10,10 +10,34 @@ public class TableIO {
     private static final XmlOutput xmlOutput = new XmlOutput();
 
     public static Table readTable(String schemaName, String tableName) throws IOException {
-        return xmlInput.readFile(FileIO.convertToFile(schemaName, tableName));
+        File schema = new File(schemaName);
+        if (!schema.exists()) {
+            throw new IllegalArgumentException("schema with name " + schemaName + " does not exist");
+        }
+        File table = FileIO.convertToFile(schemaName, tableName);
+        if (!table.exists()) {
+            throw new IllegalArgumentException("table with name" + tableName + " does not exist");
+        }
+        return xmlInput.readFile(table);
     }
 
     public static void writeTable(String schemaName, String tableName, Table table) throws IOException {
+        File schema = new File(schemaName);
+        if (!schema.exists()) {
+            throw new IllegalArgumentException("schema with name " + schemaName + " does not exist");
+        }
         xmlOutput.writeFile(FileIO.convertToFile(schemaName, tableName), table);
+    }
+
+    private static void deleteTable(String schemaName, String tableName) {
+        File schema = new File(schemaName);
+        if (!schema.exists()) {
+            throw new IllegalArgumentException("schema with name " + schemaName + " does not exist");
+        }
+        File table = FileIO.convertToFile(schemaName, tableName);
+        if (!table.exists()) {
+            throw new IllegalArgumentException("table with name" + tableName + " does not exist");
+        }
+        FileIO.deleteFile(table);
     }
 }
